Use the canonical cast names PHP 8.5 asks for
PHP 8.5 deprecates the non-canonical casts and reports every one of
them at compile time, so the notices pile up even when the code around
them never runs. Rename them to the canonical spelling: (double) to
(float), (boolean) to (bool) and (integer) to (int). Both spellings
compile to the same opcode, so the behaviour stays the same.

The sales report writes "default;" instead of "default:", which PHP 8.5
deprecates as well. Give the label its colon.

Leave the third-party code under includes/external/ and the magnalister
files untouched, since both are maintained elsewhere.

+++ admin/includes/functions/general.php
@@ -1710,7 +1710,7 @@
     $currencies = xtc_db_fetch_array($currencies_query);
 
     if ($allow_tax == '1') {
-      $price_string = (double)$price_string / ((100 + (double)$tax_rate) / 100);
+      $price_string = (float)$price_string / ((100 + (float)$tax_rate) / 100);
     }
 
     if ($price_format == '1') {
@@ -1743,7 +1743,7 @@
    * @return
    */
   function xtc_round($value, $precision) {
-    return round((double)$value, (int)$precision);
+    return round((float)$value, (int)$precision);
   }
 
   /**
